[UPD] scaffold php_task.c
